requirements.txt
python-dotenv
SQLAlchemy
pyodbc
streamlit
langchain
google-generativeai
bcrypt
